Noticed a few misconfigurations in some of the benchmarks, causing them to run on much larger datasets and misreporting the performance of the algorithms.

Before:
```
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
BenchmarkOrderOfUint32/4KiB         	 6394976	       179.9 ns/op	22768.79 MB/s
BenchmarkOrderOfUint32/256KiB       	   55310	     21739 ns/op	12058.49 MB/s
BenchmarkOrderOfUint32/2048KiB      	    4725	    245473 ns/op	8543.30 MB/s
```
After:
```
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
BenchmarkOrderOfUint32/4KiB         	24478969	        45.84 ns/op	89352.60 MB/s
BenchmarkOrderOfUint32/256KiB       	  222945	      5416 ns/op	48403.56 MB/s
BenchmarkOrderOfUint32/2048KiB      	   18451	     66004 ns/op	31773.14 MB/s
```